Release v3.4

+++ src/utils/slices/versionSlice.ts
@@ -30,7 +30,7 @@ const versionSlice = createSlice({
   extraReducers: builder => {
     builder.addCase(retriveVersionLatest.fulfilled, (state, action) => {
       if (action.payload && Constants.manifest?.version) {
-        if (parseInt(action.payload) > parseInt(Constants.manifest.version)) {
+        if (parseFloat(action.payload) > parseFloat(Constants.manifest.version)) {
           state.update = true
         }
       }
